feature: add unit test for file name generator
