+++ backend/src/controllers/image.controller.ts
@@ -1,8 +1,10 @@
+import type { Response, NextFunction } from 'express';
 import sharp from 'sharp';
-import type { Request, Response, NextFunction } from 'express';
-import generateFileName from '../util/file.util';
-import { uploadFile } from '../util/s3.util';
 import createError from 'http-errors';
+
+import generateFileName from '../util/file.util';
+import { PutObjectCommand, S3Client } from '@aws-sdk/client-s3';
+import { env } from '../config/env.config';
 
 // interface UploadRequest extends Request {
 //   file?: {
@@ -27,6 +29,25 @@ const handleUploadImage = async (req: any, res: Response, _next: NextFunction) =
       .toBuffer();
 
     const newName = `${generateFileName()}.jpg`;
+
+    const client = new S3Client({
+      region: env.AWS_REGION,
+      credentials: {
+        accessKeyId: env.AWS_ACCESS_KEY_ID,
+        secretAccessKey: env.AWS_SECRET_ACCESS_KEY,
+      },
+    });
+
+    const uploadFile = async (fileBuffer: any, fileName: any, mimetype: any) => {
+      const uploadParams = {
+        Bucket: env.AWS_BUCKET,
+        Body: fileBuffer,
+        Key: fileName,
+        ContentType: mimetype,
+      };
+
+      return await client.send(new PutObjectCommand(uploadParams));
+    };
 
     await uploadFile(buffer, newName, file.mimetype);
     res.status(200).json({ filename: newName });
