+++ backend/src/models/email.model.ts
@@ -2,6 +2,9 @@ import { z } from 'zod';
 
 import type { TypeOf } from 'zod';
 
+const emailRegex = /^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+.[a-zA-Z0-9-.]+$/;
+const phoneRegex = /^[\+]?[(]?[0-9]{3}[)]?[-\s\.]?[0-9]{3}[-\s\.]?[0-9]{4,6}$/im;
+
 export const CreateEmailSchema = z.object({
   name: z
     .string({
@@ -12,14 +15,13 @@ export const CreateEmailSchema = z.object({
     .string({
       required_error: 'Email is required',
     })
-    .regex(new RegExp('^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+.[a-zA-Z0-9-.]+$'), 'Email is invalid')
-    .min(1, 'Email too short'),
-  phone: z.string().regex(new RegExp('^(+d{1,2}s)?(?d{3})?[s.-]d{3}[s.-]d{4}$'), 'Phone is invalid').optional(),
+    .regex(emailRegex, 'Email is invalid'),
+  phone: z.string().regex(phoneRegex, 'Phone is invalid').optional(),
   message: z
     .string({
-      required_error: 'Email is required',
+      required_error: 'Message is required',
     })
-    .min(1, 'Email too short'),
+    .min(1, 'Message too short'),
   jobRole: z.string().optional(),
 });
